backend/requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
Flask-Cors

numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
joblib==1.5.3
xgboost==3.2.0

langchain==1.2.10
langchain-ollama==1.0.1
langchain-core==1.2.16
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.0.9

chromadb==1.5.2
openai==2.24.0
tiktoken==0.12.0

firebase_admin==7.2.0

huggingface_hub==1.5.0
tokenizers==0.22.2

requests==2.32.5
python-dotenv==1.2.1
tqdm==4.67.3